Include a simplistic Brick entity
Our brick entity will be used to wrap the entities that represent the
walls and background of the game play area, as well as the gray bricks
that temporarily stop the ball from moving during the game.

The game scene has a test render that displays the three different
types of bricks so that we can verify that everything works as
expected.