Include a maze entity
This is responsible for displaying the contents of the maze.

This will ened to be re-worked because it should really use the sprite
sheet to render without having to create a buttload of brick entities,
but this is good enough for the moment.

The game scene will now display the maze instead of sample bricks.