Refactor Maze and Brick
The Brick entity now derives from MazeCell, and its constructor takes
the default brick type instead of it always being a background cell.

In the Maze, there are now accessors for getting and setting the Brick
at a given maze location. These are set up to assume that a NULL at
the maze location indicates that there is just background brick at
that location.

This allows us to create a single solid and background brick, and then
use the same instance everywhere without having to have multiples of
them. In the case of the empty bricks (the background) we don't even
store those into the grid.