Update black hole sprite
This is really inaccurately named now, but it was almost completely
invisible on the empty cells, which is most certainly not what we
want.

The black part has been made white just as a quick expedient to
getting it more visible. The graphic can be tweaked more later when
it's less important.