Enhance Maze cell API
Added a new set of API methods for getting and setting the contents
of the maze. These work on MazeCell instances instead of Brick
instances, and the methods that work based on Bricks are now
implemented in terms of the new methods.

These make some upcoming changes possible while retaining nicer code
for accessing common cell types, such as bricks.