Add Maze support for the new Teleport entity
This adds the ability for the Maze to store a Teleport entity, using
the API introduced in the prior commit.

A couple of small changes are needed to the Maze class for this to
work:

Firstly, when rendering we need to make sure that we render every cell
that is not already a brick with the background brick behind it first
(since in this case the entity is probably transparent).

Secondly, we need to overrride the update() method in the Maze entity
to also call the update() method for the Teleport entity, since it is
animated and it's animation won't be driven otherwise.