Include an Arrow entity
Like the Brick entity, this wraps both kinds of arrows into a single
entity.

The Arrow knows what type it is (which currently only determines what
color it renders itself as), and what direction it thinks that it is
facing.

The sprite sheet has animations set up so that it can idle facing
either direction or animate rotating itself over to the new facing
when it changes directions.