Animate Gray bricks
Set up the Brick entity so that the Gray brick can animate itself
vanishing out of the level (and appearing, since that is essentially
free and might make level generation look nicer).

In order to make the rendering code easier on the eye, the sprite
sheet for the gray brick (and also the score tile, which is going to
be  added next) has been modified to have a solid background that
represents the empty brick, so that we don't have to double-render it.