Include EntityPool class
This is a quick rough take on the concept. Basically the idea is that
this is generic and should be part of the engine, although for now
we'll leave it as is.

It's basically a glorified array that knows how to invoke the update
method for all stored, live entities. It also has the ability to mark
stored entities as either dead or alive, so that dead ones do not
update.

I think this should actually be an ActorPool and not an EntityPool if
it's in the engine; plus it should probably use a generic constructor
so that it can pull items out nicer. We'll leave that for another
day, though.