Enhance the Arrow entity
The Arrow entity is enhanced so that the type and direction of the
arrow can be directly modified after creation. Previously the type was
set permanently and the direction could only be set by animating a
flip to the new direction.

Additionally, the type and direction parameters on the constructor have
defaults set (although they are probably not actually needed at all)
to facilitate creating a bunch of them all at once easier.