Get ready for maze generation
In preparation for actually generating the maze, we do a littlqe
cleanup and refactoring here.

Firstly, the bogus entries are no longer added to the maze at
generation time, and they're not changed by the update method.

Additionally reset has been refactored out a bit so that what it
actualy does is call a method that erases the maze, so we can
compartmentalize everything better.

Lastly we create a few helper methods that we can use during the
generation process.