Generate arrows into the maze
For arrows, we generate a certain number of arrows per row instead of
randomly selecting positions. This seems to make a more visually
appealing maze to my untrained, biased eye.

This uses the new EntityPool class to hold all of the arrows so that
we can keep a fixed number of them around, and easily update all of
them as needed.