Swap EntityPool for ActorPool
I couldn't help but do research (during my morning walk of all times)
to see if TypeScript supports generics in the way that I thought it
did, and it surely does.

So, as a first step, lets change the name and signatures of the class
to properly know that it's actually Actors that are at the base of
the hierarchy and not Entities.