Tweak arrow generation
Modify how arrows generate slightly so that they will not generate in
the tile directly under a black hole. This makes a more visually
pleasing maze and reduces (but not not eliminate) the chance that a
ball can be pushed by an arrow into a black hole without it dropping
first.