Extend the ActorPool to be able to kill all actors
For the purposes of being able to use the actor pool to reuse objects
it seems like a good idea to be able to kill all of the objects that
are currently alive in it.

This is used in the Maze whenever the reset() method is called in
order to generate a maze, to ensure that we don't keep calling update
for entities that are not on the screen.