Work around an engine "Feature"
The engine will throw an exception if you try to add an image preload
after the game has started. Turns out that it also incorrectly tries
to add a preload if you create an entity that loads an image after the
game as started.

The proper fix is to either fix the engine or (better) have each
entity class have a class static preload of the image and use that
instead, but since I'm not sure what the ramifications of that might
be, we hack around the problem by pre-filling the list of entities of
each type before things kick off.