Generate balls into the level
This currently only generates balls for the player into the grid. It
should actually also have a second data structure that represents a
hidden version of the top row that contains the balls that the
computer will be using.

A key would be used to see what the computer's ball layout looks like
while you are taking your turn.