Allow Maze to handle a mouse click
The Maze entity now has a method that can be invoked with a position
inside of itself that it will treat as a mouse click.

This will be used for debugging by allowing us to interact with the
maze on click. Currently it's set up to only try and handle clicks on
balls in the top row, although that is subject to change.