Fix a generation bug
The generation of the bonus tiles was accidentally pulling the
entities out of the gray brick actor pool instead of the bonus brick
pool.

Although the generation specified that they play the animation for an
appearing bonus brick which made them look correct, they were actually
considered to be gray bricks by type, which is wrong.