Improve the click handler in the maze
This is now clever enough that if you click on a gray brick or a bonus
brick it will vanish (or appear if you click where it is and it is
currently vanished).

Clicking on an arrow will flip that arrow over.

Lastly, clicking on a ball will vanish it, but only if it is in the
top row. This is a place holder for the logic by which we will start
the ball dropping.