Include a Marker entity
In order to work on dropping the ball through the maze, we need it to
work two ways: one step at a time, as well as capturing an entire path
all at once (for use in AI).

In order to test this out we're going to start with calculating all of
the steps at once, and so we are implementing a new Marker entity.

This does nothing but render it's bounds for debugging. We will use it
to mark what path the ball takes on it's path.