Include Marker handling in the Maze
Clicking in an empty spot in the maze will cause a marker to be placed
there, while clicking on a marker will remove it. Additionaly, when
a ball is clicked on, all markers are removed from the maze first.

The Maze has an ActorPool for handling markers, which starts out empty
(since we need preload to be finished before we can create them), and
a single method for fetching/creating one.