Include the start of ball dropping logic
This is crude and extremely verbose (and not fully functional), but it
is a proof of concept.

When a ball is clicked on, the click handler will repeatedly invoke a
function given what it thinks the current position of the ball is, to
see where the ball would end up. This stops when the ball hits a
location from which it can no longer move.

We pass through bonus blocks and vanished gray bricks (but not
visible) ones, but we are not capable of teleporting just yet.

Additionally, the arrows don't flip like they're supposed to as we
pass through them; this is currently by design, since this method
would be used by the AI and needs to calculate where the ball would go
if it dropped right this second, without actually changing anything
(except that for debug purposes we kind of are).