Make Marker and Teleport not block the ball
A Marker should not block the ball from moving because it doesn't
actually represent level geometry; it's just there so we can track
where the ball would have gone.

We are temporarily marking the Teleport as not blocking the ball so
that the ball will pass through it instead of stopping, which is not
overly helpful for testing.