Modify how Markers work
The initial implementation of Markers just inserted them directly into
the maze at the provided location, which clobbers the maze contents at
that location.

The Maze entity now has a separate container for holding whether the
contents of the maze at a particular location is marked or not, and
this is queried at render time.

This means that we can mark a location without clobbering it, so that
we get a truer representation of where the ball moved.