Modify how touch works in MazeCell
The logic of how the new touch method works has been switched to be
deferred.

Previously, it was triggered on the step where the ball enters the
cell. Now it is instead triggered at the start of the move, after the
ball has already moved on top.

In addition, the method now takes a Point that represents the location
that the ball is at while touching, and this point can be modified.

In the movement handler, when we invoke the method we check to see if
it modified the point, and if it did, we treat that as the movement.
In this case, the new position is treated as always valid.