Update to the update; no touch after touch
Here's an issue I didn't think about until after the last change:

The start of the movement code checks to see if the ball is on top of
an entity that can change the position, and if it is, it changes the
location. What happens if this change drops the ball onto a cell that
has an entity that can also do this?

The answer is that on the next move cycle, the ball immediately jumps
to a new location without dropping. potentially. This is of concern
with the black hole entities, where the position they jump the ball to
is another black hole.

This ultimately means that once the ball hits a teleport, it moves
forever.

To get around this, we insitute a rule that back to back movement
calls are not allowed to trigger a ball movement.