Support Teleporting
This adds the ability of the Teleport entity to contain a list of
potential destination locations and to select one in response to the
ball touching it.

When a maze generation starts, all known teleport locations are
erased, and then as the new ones are added, they are added back.

This is enough (along with previous features) to support the ball
teleporting to another location.

Currently this mechanism is unsafe in that the code that moves the
ball doesn't check to make sure the destination location is blocked,
which  it might be if a ball is blocking up the exit selected.