Clean up the Marker entity
This was originally created to take it's size from a maze, but we can
no longer rely on the maze cell size because it's not defined what
order the callbacks for preload being finished are called in.

Instead of giving the constructor a redundant maze that it ignores,
followed by us manually setting the dimensions, we might as well just
set those dimensions ourselves and cut out the middle man.

Also some of the comments show that I lifted the code right from the
Arrow entity.