MazeCell refactor
For upcoming changes, the API methods in MazeCell that are used to
move the ball have been renamed and had their signatures modified. The
rename is just to make the code more readable and to find errors
faster.

changeBallLocation() has become ballCollsion(), and gets the maze, ball
and location of where the collision is happening. Instead of modifying
the Point given if it needs to modify the location, it returns either
null or a new location.

didChangeDirection() has become didMoveBall(), and gets the ball that
it moved.

touchingBall() has become ballTouch(), and like ballCollision() gets
the maze, ball and location of the touch.

At the current time the old API is still there and in use.