Allow the Ball entity to track movement
In order for the next features to be implemented, the Ball entity can
now track exactly how it is moving through the maze by recording its
changes in direction.

A property is exposed to allow us to get and set what the last
movement type was for any given ball. By tracking the staTe of this
value we know if the ball arrived where it currently got due to
gravity dropping it, a teleporter, and so on.