Arrow and Ball track Ball movement now
Using the new tracking API in the Ball entity for knowing what
direction it's moving, the arrow reacts to being told that it
successfully  moved the ball by setting the move direction of the ball
as appropriate.

The Ball entity now tries to move the ball left or right if it was
already going left or right, mimicking what the Arrow entity would do.

This allows the ball to keep rolling over balls that are below it;
this adds a bit of a strategic element to the game.