Include debug command: Toggle
This will toggle an arrow or a brick between it's two states. For an
arrow this means transforming between a normal and an automatic arrow,
while for a brick it means toggling between a bonus and a gray brick.

In the latter case the number of objects is predefined, so the toggle
can only happen if there is a brick of the opposite type available
in its pool.