Include debug command: Add brick
This will insert a brick entity into the maze at the current debug
location, as long as it is empty.

This needs to pull a brick from a pool in order to add one. It will
try first from the gray brick pool, then from the bonus brick pool if
the gray pool is empty.