One last debug option
Modify the left click handling code so that the only part of it that
will trigger in non-debug mode is dropping a ball. Everything else
like toggling a marker, flipping an arrow of vanish/appearing a brick
is now a level editing feature only.

Additionally, slight tweaks to marker code. Clicking on a Black hole
can set or clear a mark on this cell. Clicking on a brick that is
marked will cause the mark to be removed while the brick is
vanishing/appearing.

This allows for the full removal of any markers left by the ball
dropping code on otherwise unclickable entities, which was driving the
OCD part of my brain nuts.