Include ball dropping
This is a bit crude at the moment, but it works.

Essentially what we do is remove the ball from the grid and hold it in
a separate variable, which we then track for movement. By keeping track
of the tick at the last time the ball moved and the movement between
ticks, we can tell when it's time to invoke the method to skip to the
next position in the grid.