Clean up rendering
Now that we're actually into the meat of this thing, tidy up the
rendering code a little bit.

Instead of performing a bunch of math for every blit of every cell
to determine the correct position, change it just some additions.
This is faster PLUS looks nicer.

The blitting code outside of the grid loop aliases the points it's
about to use just for code brevity.