Clean up rendering a bit
We're on track to have the rendering code not require everything that is
rendered be stored in the maze contents. Along the way is this temporary
step to keep everything working as expected.

What we have done is removed from Maze the check that causes the code
to render a background tile underneath entities that are not bricks
so that they appear properly.

Instead of that, MazeCell now has a boolean flag per instance that
indicates if this instance has transparency, and the native render
method will do that on it's own.

This removes some cruft from the rendering code (more to come).