Rendering cleanup; redact previous change
For the next step in the rendering process we have a method that will
render an entire background of a maze (including the bounding walls)
before we do anything else.

This means that we (eventually) don't have to actually loop over the
maze contents to render it (except for markers) although we currently
still are doing that.