More rendering refactoring
We have added to the ActorPool the ability to invoke the render method
on all actors that are still live based on their current position on
the screen.

Now the maze renders almost everything by first rendering a
background, then rendering all live entities from pools, and then
finally markers.

There is still a little bit to be done here; teleports do not render
yet (they are not in a pool) and the dropping ball is a little visually
wonky because the ball dropping code does not properly update the
screen position as the ball is in transit.