Fix moving ball rendering
Fix the issue where while the ball is moving if we don't manually draw
it i, it doesn't render in the proper location. This was an issue only
because as the ball is dropping, it's screen location is not being
properly calculated.

This is now fixed, so the manual render is gone.