Change render order of teleports
Since it is possible for a ball to end up sitting on top of a teleport
entity unable to move (due to teleporing in and being blocked from
further travel) we swap the render order of things.

Now teleports render first and balls render after them, so that if
they  overlay you can see the ball and the faint motion of the black
hole behind it, instead of just seeing a black hole tha is vaguely
ball colored in the rotating gap.