Use new API to include entity reaping
Implement a new API in the Maze entity that can scan all of the live
entities in an ActorPool of actors that implement the new interface,
and find all which are currently hidden and remove them from the maze.

This cleans up the maze content and allows us to stop trying to
rendering entities which are no longer needed.

For now this only operates over the Ball entities, which means that
once they reach the bottom of the grid, they vanish and then are
destroyed once the ball finishes vanishing.