Implement move tracking and gray brick removal
Every time a ball has finished moving and either come to rest or
finished vanishing away, we check to see if there are any balls still
left to play.

If there are none, we tell all of the gray bricks to vanish. Once this
happens any balls that move and come to a rest no longer perform this
check.

All that is left now is to perform the loop whereby we start at the
bottom right side of the grid and slowly iterate back up and left
dropping balls and vanishing them when they come to rest, which would
signal the end of the round.