Clean up debug code a bit
This is just some simple cleanup. Since bricks that are vanished out
of the maze are reaped now, the code for unhiding them is no longer
needed since it can't do anything anyway.

This also removes the control from the scene and the documentation.

There are a couple of smaller changes as well in a similar vein.