Drop balls automatically at game end
This new logic detects when all of the gray bricks have finished
vanishing, and then one by one is selects balls from the maze and
drops them one last time.

The same drop code is used as for the regular drop, except that in
this phase the ball is always vanished when it stops moving, even if
it doesn't hit the goal line.

We move balls from the bottom up so that every ball gets a fair chance
to move through the maze without being blocked.