Create new MazeContent class
This class wraps the maze contents and the operations for getting and
setting cells and markers in it, removing that from the Maze itself.

This cleans up the code and also allows us to  do something like pass
the content easily as a get() property so that e.g. the scene that owns
us can modify, which means we can also get a bunch of the debug code
out of there as well.