Add ActorPool tracking to MazeCell
This adds to ActorPool an interface that extends the Actor class and
adds a single optional parameter named "pool" that is an ActorPool.
The ActorPool generic description now says that it takes an object of
this new interface, which means it works with any Actor which may
optionally have such a  property.

The MazeCell entity adds this property, which means that all
subclasses get it as well.

The idea here is that any entity that is stored in an ActorPool tracks
its pool, which means that we can use any entity to remove it from the
associated pool (if desired) without having to know what type of
object that is is.

This is defined in this manner (optional parameter) so that we don't
have to update the Actor class in ts-game-engine to provide this
property by default. This will happen when this class gets rolled into
ts-game-engine.