Remove almost all of the instanceof uses
We now favor using the internal name of the entity, which for our
purposes here is the name provided to the MazeCell constructor.

ts-game-engine uses this for debugging so it's exposed, but it
probably should be. This has been added to the new interface to
remind us that it should be added along with everything else.