Refactor most debugging code into a new class
This does for debugging what the previous commits did for the
generation of the maze. This pulls all of the debug options out and
puts them into their own class.

This still needs a bit of work, si\nce the flag and a couple of other
things are still held in the Maze.