Clean up debugging calls
Shift the onus of checking to see if debugging is turned on or not
from the calling class to the MazeDebugger class, which leaves without
doing anything if debugging is not turned on.

A boolean return tells you if anything was done or not, so that the
Scene can eat keys as needed.