Include walking
The Scene can now make the player walk left and right in the scene. If
the player is not facing in the correct direction, it rotates to face
the given direciton, otherwise it walks in that direction.

The movement is constrained to the limits of the columns of the maze
grid between the walls.