Set up for AI; include simulation flag
In order to use our existing code for regular and AI purposes, we now
are able to pass a parameter to the MazeCell methods that are
responsible for touching and redirecting the ball, so that the method
can tell for what purpose this is happening (regular or AI).

The existing code invokes this with a parameter that indicates that
this operation is a regular one so that everything happens as per
usual.