Implement simulation save/restore for Ball
For the ball the only state we need to track is the position we were
at before we were pushed (so we can go back there) and the score that
is associated with this ball.

Calling code is currently expected not to hide us visually if we're in
a simulation.