Implement simulation save/restore for Arrow
For the arrow the only state we save is the direction we're pointing.
This also updates the flip method to take a boolean that indicates if
this flip is part of a simulation, so that it knows if it should
change the visual orientation to match or not.

Currently the simulation code does not invoke the update() method, so
it is basically operating under the hope that an automatic arrow will
be pointing the right direction when the ball gets there (same as a
human).