Implement simulation save/restore for Brick
For a bonus brick, we handle touches different. If we are not
simulating, we do what we always did (only nicer; forgot to refactor
some code).

If we ARE simulating, we update the score in the ball that touched us
to indicate that we actually were collected by this ball and then set
a flag saying it happened so we won't do it again.