Implement simple AI tests
This just has the AI run to select a ball, then jump the player there
and push it so we can see what happens.

This updates the player to have a new test method for jumping to a
specific location in the maze instead of just going left and right.