Add computer animations to Player entity
This is a simple operation but allows us to support the computer
player as well. This adds in the animations that are needed in order
to display as the computer and a parameter in the contructor to
specify which type of player to be.

The code to select the appropriate animation was already set up to
work as desired so no changes were needed.