Enhance Player entity movement API
This is how this should have been implemented originally, but I just
threw a quick couple of functions together for expediency to try and
get the change finished before the end of that devember day.

This merges the methods for moving left and right to simple
translation and updates the documentation comment on this and the
jumpTo method to not suck.