Fix AI bug
During simulation of a drop, a vanished gray brick would count as a
bonus brick and add points to the ball, even though in practice this
is not a situation in which points are scored.

As a result of this the AI would prioritize moves through vanished
gray bricks over regular drops for no discernible reason.