Enhance AI
This is a feature of the AI that I wanted to implement originally but
which I did not get around to in the press for getting the basic
concept up and running.

Essentially, when the AI is calculating the move that it wants to
make, it  should pretend that the gray bricks are not actually there
(even if they are) under the theory that when the round ends, such
balls will drop  and would theoretically score more.

To this end, the collision API for detecting if an entity blocks a
ball now gets the isSimulation parameter so that this can happen.