Include new StateMachine class
This is a first pass and is subject to changes, so we'll see what
happens.

This is set up so that the machine maintains both a current and a
previous state (i.e. it has a state stack size of 1), allowing us to
not only track our current state but also make rudimentary decisions
about where to go next.

There is a change listener interface that allows the machine to
directly notify interested parties when the state changes so that they
can take the appropriate action.

This code currently has a proviso (in comments only) that warns again
changing the state of the machine while inside the change handler, but
I think as long as there is only ever a single listener this is safe.
In the case of a multiple listener, notifications to objects that get
notifed after the current cycle might be undefined.