Add state machine to Game scene
This adds in the actual property that includes the state machine in
the game state. We set up a state and priorState property right on the
Scene class instance itself so that we can get at the value easier.

This checks at the time that the scene is activated to see if it is
currently no state (startup) and if so, changes the state to the level
generation state, which will then kick off the level generation.

This is basically a roundabout way to do what we were already doing.