Constrain player input to provided state
This shifts the code responsible for allowing the player to take a
turn (or do something that can happen only on their turn) into its own
method, which will handle it as before but do nothing if the state
does not indicate that this is the player's turn.

This proves that the player will not be able to interact with anything
when it is not their turn.