Respond to ball drop finishing properly (moreso)
Instead of just always being the turn of the player, we now take turns
by swapping to an appropriate state depending on what the prior state
was.

This is only a base implementation; it doesn't know where to go after
the final ball drop, it doesn't swap balls for the computer or player,
and so on.