Use state logic to make a computer take a turn
This introduces the logic required to have the computer player take a
turn. When this happens it will select a move, then walk to that
location and push the appropriate ball.

This lacks the change to the state where we know that the end of the
round is at hand, so when this happens the two players just keep
swapping back and forth doing nothing.