Enhance MazeContent class
We now have two separate arrays that store the playable balls for each
player, and have the ability to swap what set is currently visible in
the top row of the maze.

We can also query to see if a given player has a move or not, allowing
us to skip the turn of a player if needed.