Fix bug in MazeContents
In favor of pasing a parameter to the swap function that swaps balls
between the computer and the player that says if it should save the
ball state or not, instead allow setting the visible ball state to
clobber the existing ball row.

This will be futher enhanced in the immediate future, but this causes
a problem in that we want the player balls to be visible first, but
because that is the default, the swap function swaps them and displays
the wrong one.