Update Maze end of game tracking
Now that we have a separate set of balls for the compute and for the
player, we need better checking to determine if all balls have been
played or not (since not all are visible).

There is a method we can use for this, which makes the existing code
much simpler.