Fix a bug; do not reap balls that are hidden
Yesterday's woes are caused by the fact that the update loop is
constantly trying to reap hidden balls by marking them as dead.

This is an issue because a dead ball will not render or update, which
means that any computer balls that are hidden are almost immediately
killed, causing them to no longer update even though they're still in
the maze.

Since we render all balls that are alive, this means that once the
scene is set up with a set of live balls, it always looks like this
even when the balls are swapped into the maze, because they won't
overwrite themselves. Grrr.