Re-enable the ability of the computer to take turn
Now that we have fixed the underlying issue, we can proceed with
allowing the computer to take a turn again when it becomes its turn,
since the balls are actually there now.

The rest of the code should work as we want it to from this point, but
I'll check anyway.