Add a "Just in case" check
Just to ensure that nothing goes away, before we hide a ball while
swapping we first remove it from the maze. This makes sure that the
update loop can't hose it out from under us.

I'm pretty sure this can't actually happen but I'm not 100% sure in
what sequence the code runs if it is responding to input, so bette
safe than sorry.