Check that a play is possible before switching
We can now check to ensure that if it is to be a player's turn, that
there is a valid move for that player first. If not, we instead skip
to the next player's turn.

In the case that neither player has a valid move, we transition to the
state that handles the final ball drop, although currently that's not
how this state is handled.