Switch to removing gray bricks when no moves left
Now instead of swapping to the final ball drop state, we swap to the
state in which we remove all gray bricks. We handle this state by
using the state timer to slowly remove all of the gray bricks (telling
them to vanish).

Currently the code that reaps vanishing bricks is just spewing a
console log every time it knows all of the gray bricks are gone.