Swap to final ball drop state when bricks gone
Now the maze will invoke a method on the listener to tell it that all
of the gray bricks have vanished out of the maze and have been reaped.

The Game state uses this to transition into the final ball drop state,
so that it can start the final ball drop of the round.