Put final ball drop under the control of states
This puts the final ball drop into the control of the state machine;
when we are in the final ball drop state, we select a ball and drop
it, and then return to this state once the ball is done.

There is currently a bug in this in which sometimes during the final
ball drop a ball (that is blocked?) starts dropping, but the flag that
says that it is done moving is not set.