Allow for the notion of no gray bricks
This is probably only of use when debugging, but as originally written
the code would get stuck and not be able to enter the final ball drop
if there are no gray bricks in the level because in that case there
are none to reap.

This is fixed by having the method that would normally remove such a
brick return false if there are none left to remove; when this happens
the code can immediately jump to the next phase.