Fix final ball drop bug
It turns out that the reason for this failure was entirely because I
incorrectly stopped the search for balls to remove for being vanished
one column short of the edge.

This means that any final ball dropped in the last column of the maze
will vanish but not be reaped. Thus the flag for the ball being done
will never get set and everything stalls out.

Whoops.