Include method to find and vanish blocked balls.
This will be used during end game, just prior to vanishing all of the
gray bricks and dropping the final balls.

We want to be able to find all balls that have no chance of moving
because they are either stopped on an arrow, or stopped on top of a
ball waiting on an arrow, and then remove them away like the gray
bricks do.