Implement ball removal state
This hooks into the previously created method and will continuously
try to remove blocked balls from the maze until the method tells it
that there are no further balls to remove and all that were removed
are now done with.

We skip to this state when all plays are exhausted and then to the
remove gray bricks state once this is completed.