Fix bugs in blocked ball removal
This code went through several iterations, and so some things were not
in the correct order (e.g. null test AFTER other tests).

There was also an issue where a vanished ball would trigger the  ball
drop complete method and then throw an error because there is no
dropped ball, which I have currently patched away but this needs to be
handled better.