Fix drop ball bug introduced yesterday
Yesterday I introduced some new code where we could detect a vanished
ball but not trigger the notification that a ball drop was complete if
the ball was being vanished due to it being a blocked ball.

This was a bit flawed but unfortunatey I didn't notice it until it was
too late. Basically it would not trigger for a ball that vanished
because it got to the goal, because I used the wrong variable to track
if the event should trigger or not (_droppedBall instead of
_lastDroppedBall) and since _droppedBall gets cleared as soon as the
vanish is called, such a ball is treated like a blocked ball.

This has been fixed by using the correct variable.