Tweak black hole generation
We now generate the black holes into a narrower band in the maze
content area; they will not populate near the top or bottom of the
grid, and they won't spawn next to the walls.

Additionally, the range between them has been extended a tile and we
ensure that there is only ever a single teleport per column at most,
to maximize the chaos of the ball jumping around in the maze.