Enhance Teleport; select non-blocked destination
This change adds to the Teleport entity the ability to ensure that the
position that it returns in response to ballTouch() is a location to
which the ball can be transported without blocking.

This is done by conducting a pre-check to ensure that there is at
least one valid destination to select, so that it won't loop forever
if all of the destinations are blocked.