Unplug blocked exits to Teleport entities
It is possible for a ball to exit a Teleport and then not be able to
drop further. When this happens, the ball movement stops and the ball
gets put back into the maze at the location of the Teleport exit.

This clobbers the Teleport out of the maze at this location, although
it still visually renders at this location because the render is based
on destinations (they all share the same entity). This results in that
being an "exit only" teleport; the ball can't enter the teleport there
because it's not in the maze any longer.

Now when we determine that we are about to drop the ball down, we
check to see if the position that we're currently occupying is empty
and also a destination in the single teleport entity.

If this is the case, we can re-insert the teleport, which allows this
location to be the source of a teleport operation again.