Include values for tracking rounds
This introduces a current round and maximum round to the game state,
along with methods for resetting the game to a new start for a given
number of levels, advancing levels, and seeing if it is the end of the
game or not.

This works such that a maximum level count of 0 or smaller counts as a
single round game played with half balls. Any game that is more than a
single round will report the last round condition when it's the last
round, which can be used to turn on the automatic arrows in maze
generation.